pyproject.toml
[project]
name = "pfloss"
version = "0.1.0"
description = ""
authors = [
    {name = "PGuin",email = "138515193+pguin-sudo@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "pfloss", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
